rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.340
langchain-community==0.0.8
openai==1.3.0
pymupdf==1.23.8
numpy==1.26.0
pandas==2.1.1
pydantic==2.4.2
